requirements.txt
alabaster==0.7.12
argh==0.26.2
asn1crypto==1.3.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cmake==3.16.3
cryptography==2.8
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
dlib==19.19.0
docutils==0.16
entrypoints==0.3
face-recognition==1.2.3
face-recognition-models==0.3.0
flake8==3.7.9
future==0.18.2
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
imutils==0.5.3
inflect==4.1.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
iso8601==0.1.12
isort==4.3.21
jaraco.itertools==5.0.0
jedi==0.14.1
jeepney==0.4.2
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
keyring==21.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.4
numpy==1.18.1
numpydoc==0.9.2
opencv-python==4.2.0.32
packaging==20.1
pandocfilters==1.4.2
parso==0.6.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prompt-toolkit==3.0.3
psutil==5.6.7
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.11.0
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
pyxdg==0.26
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
regex==2020.1.8
requests==2.22.0
rope==0.16.0
Rtree==0.9.3
scikit-learn==0.22.1
scipy==1.4.1
SecretStorage==3.1.2
six==1.14.0
sklearn==0.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==4.0.1
spyder-kernels==1.8.1
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
ujson==1.35
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
wrapt==1.11.2
wurlitzer==2.0.0
yapf==0.28.0
zipp==2.1.0
